requirements.txt
openai>=1.40.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0
matplotlib>=3.7.0
plotly>=5.15.0
openpyxl>=3.1.0
pyarrow>=12.0.0



